% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsetools-package.R
\docType{package}
\name{rsetools-package}
\alias{rsetools}
\alias{rsetools-package}
\title{rsetools: Easier Research Software Engineering}
\description{
The goal of rsetools is to make research software engineering easier.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/maurolepore/rsetools}
  \item \url{https://maurolepore.github.io/rsetools/}
  \item Report bugs at \url{https://github.com/maurolepore/rsetools/issues}
}

}
\author{
\strong{Maintainer}: Mauro Lepore \email{maurolepore@gmail.com} (\href{https://orcid.org/0000-0002-1986-7988}{ORCID})

}
\keyword{internal}
