% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_dicitionary.R
\name{use_dictionary}
\alias{use_dictionary}
\title{Draft a data dictionary}
\usage{
use_dictionary(data)
}
\arguments{
\item{data}{A data frame.}
}
\value{
A \link[tibble:tibble]{tibble::tibble}.
}
\description{
Draft a data dictionary
}
\examples{
use_dictionary(mtcars)
}
