% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_pr_checklist.R
\name{use_pr_checklist}
\alias{use_pr_checklist}
\title{Use stuff}
\usage{
use_pr_checklist(
  repo_spec = "maurolepore/rsetools",
  path = ".github/pull_request_template.md"
)
}
\arguments{
\item{repo_spec}{A string identifying the GitHub repo or, alternatively, a
GitHub file URL. Acceptable forms:
\itemize{
\item Plain \code{OWNER/REPO} spec
\item A blob URL, such as \code{"https://github.com/OWNER/REPO/blob/REF/path/to/some/file"}
\item A raw URL, such as \code{"https://raw.githubusercontent.com/OWNER/REPO/REF/path/to/some/file"}
}

In the case of a URL, the \code{path}, \code{ref}, and \code{host} are extracted from it, in
addition to the \code{repo_spec}.}

\item{path}{Path of file to copy, relative to the GitHub repo it lives in.
This is extracted from \code{repo_spec} when user provides a URL.}
}
\value{
See the corresponding function in
\href{https://usethis.r-lib.org/reference/index.html}{usethis}.
}
\description{
Gets the content of a file from GitHub, from any repo the user can read, and
writes it into the active project. This function wraps an endpoint of the
GitHub API which supports specifying a target reference (i.e. branch, tag,
or commit) and which follows symlinks.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
use_pr_checklist()
\dontshow{\}) # examplesIf}
}
